+++ wger/nutrition/apps.py
@@ -13,12 +13,14 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU Affero General Public License
-# along with Workout Manager.  If not, see <http://www.gnu.org/licenses/>.
+
+# Django
+from django.apps import AppConfig
 
 
-# wger
-from wger import get_version
+class NutritionConfig(AppConfig):
+    name = 'wger.nutrition'
+    verbose_name = "Nutrition"
 
-
-VERSION = get_version()
-default_app_config = 'wger.nutrition.apps.NutritionConfig'
+    def ready(self):
+        import wger.nutrition.signals